requirements.txt
environs==9.5.0
geopy==2.2.0
python-telegram-bot==13.12
redis==4.3.3
requests==2.28.0